feat: automate Whatsapp message on status change
After a registration/unregistration occurs if the room happens to change
status a POST request to https://whapi.cloud (the official API by Meta
is too restrictive not allowing to send messages to Channels) is made to
send a message with the status update on the room Whatsapp channel.
This requires two new environment variables to be defined
* WHAP_CHAT_ID: the ID for the Channel
* WHAP_TOKEN: the access token for whapi.cloud
which in my case are passed via AWS Labmda secrets.